+++ R/stepwise.R
@@ -85,9 +85,9 @@
 #'
 #' @param best_n Maximum number of models to retain for each variable count (default: 3)
 #' 
-#' @param test_ratio Ratio of training set to test set (default: 0)
+#' @param test_ratio Proportion of the dataset allocated for testing (e.g., 0.3, which means 30\% of the dataset is used for testing), with the remainder reserved for training, enabling train-test validation.
 #' 
-#' @param feature_ratio Ratio of feature selection to full model (default: 1), this is only valid when strategy is "forward".
+#' @param feature_ratio Proportion of candidate features sampled uniformly at random during forward selection (default = 1). This randomized selection helps identify the best variables while reducing the risk of overfitting, and is only valid when strategy is "forward".
 #' 
 #' @param seed Seed for random number generation (default: 123), this is only valid when test_ratio or feature_ratio is specified.
 #'